fix: ajustes solicitados no code review
